+++ apps/meteor/app/api/server/v1/channels.ts
@@ -54,7 +54,6 @@ import { normalizeMessagesForUser } from '../../../utils/server/lib/normalizeMes
 import { API } from '../api';
 import { addUserToFileObj } from '../helpers/addUserToFileObj';
 import { composeRoomWithLastMessage } from '../helpers/composeRoomWithLastMessage';
-import { getLoggedInUser } from '../helpers/getLoggedInUser';
 import { getPaginationItems } from '../helpers/getPaginationItems';
 import { getUserFromParams, getUserListFromParams } from '../helpers/getUserFromParams';
 
@@ -310,7 +309,7 @@ API.v1.addRoute(
 				rid: findResult._id,
 				...parseIds(mentionIds, 'mentions._id'),
 				...parseIds(starredIds, 'starred._id'),
-				...(pinned && pinned.toLowerCase() === 'true' ? { pinned: true } : {}),
+				...(pinned?.toLowerCase() === 'true' ? { pinned: true } : {}),
 				_hidden: { $ne: true },
 			};
 
@@ -493,7 +492,7 @@ API.v1.addRoute(
 				checkedArchived: false,
 			});
 
-			await eraseRoom(room._id, this.userId);
+			await eraseRoom(room._id, this.user);
 
 			return API.v1.success();
 		},
@@ -784,7 +783,6 @@ API.v1.addRoute(
 				const teamMembers = [];
 
 				for (const team of teams) {
-					// eslint-disable-next-line no-await-in-loop
 					const { records: members } = await Team.members(this.userId, team._id, canSeeAllTeams, {
 						offset: 0,
 						count: Number.MAX_SAFE_INTEGER,
@@ -1147,9 +1145,7 @@ API.v1.addRoute(
 				return API.v1.failure('Channel does not exists');
 			}
 
-			const user = await getLoggedInUser(this.request);
-
-			if (!room || !user || !(await canAccessRoomAsync(room, user))) {
+			if (!(await canAccessRoomAsync(room, this.user))) {
 				throw new Meteor.Error('error-not-allowed', 'Not Allowed');
 			}