+++ packages/apps-engine/deno-runtime/handlers/app/handleOnEnable.ts
@@ -2,8 +2,10 @@ import type { App } from '@rocket.chat/apps-engine/definition/App.ts';
 
 import { AppObjectRegistry } from '../../AppObjectRegistry.ts';
 import { AppAccessorsInstance } from '../../lib/accessors/mod.ts';
+import { RequestContext } from '../../lib/requestContext.ts';
+import { wrapAppForRequest } from '../../lib/wrapAppForRequest.ts';
 
-export default function handleOnEnable(): Promise<boolean> {
+export default function handleOnEnable(request: RequestContext): Promise<boolean> {
 	const app = AppObjectRegistry.get<App>('app');
 
 	if (typeof app?.onEnable !== 'function') {
@@ -12,5 +14,9 @@ export default function handleOnEnable(): Promise<boolean> {
 		});
 	}
 
-	return app.onEnable(AppAccessorsInstance.getEnvironmentRead(), AppAccessorsInstance.getConfigurationModify());
+	return app.onEnable.call(
+		wrapAppForRequest(app, request),
+		AppAccessorsInstance.getEnvironmentRead(),
+		AppAccessorsInstance.getConfigurationModify()
+	);
 }